+++ collective/venue/eventaccessor.py
@@ -32,12 +32,12 @@ class VenueEventAccessor(EventAccessor):
         location_ref = ILocation(context)
         location_uid = location_ref.location_uid
         location_notes = location_ref.location_notes
-        location_url = location_ref.location_url
         location = uuidToObject(location_uid)
 
         meta_basic = IBasic(location, None)
         add = IAddress(location, None)
 
+        location_url = None
         ret = u''
         if meta_basic and add:
             # I'm a location reference.